% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDocChangeEntry.R
\name{addDocChangeEntry}
\alias{addDocChangeEntry}
\title{Add entry to a YAML document change list}
\usage{
addDocChangeEntry(x_y = "0-0", authors = "", modification = "",
  status = "", doc_change_yaml = "config/document_change_record.yaml",
  simple = T)
}
\arguments{
\item{authors}{The authors that made changes}

\item{status}{version statius, i.e., Draft, Private, Public}

\item{doc_change_yaml}{The yaml file with document change metadata}

\item{simple}{Use simple version number}

\item{x-y}{The major version}

\item{modifications}{Comment on version/modifications}
}
\value{
If the last entry of the document change metadata is different than 
the present version the YAML metadat is updated with a new entry.
}
\description{
Versioning follows this specification http://programmers.stackexchange.com/questions/141973/how-do-you-achieve-a-numeric-versioning-scheme-with-git
}

