% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDocChangeTab.R
\name{addDocChangeTab}
\alias{addDocChangeTab}
\title{Add a document change record table}
\usage{
addDocChangeTab(yaml_path, table_out = "kable")
}
\arguments{
\item{yaml_path}{Path to YAML file}

\item{table_out}{If option 'kable' given output is a markdown table, otherwise a dataframe}
}
\value{
If option 'kable' given output is a markdown table, otherwise a dataframe
}
\description{
Convert a YAML document listing 'document changes' into a markdown table
ready for inserting in a document.
}
\examples{
yaml_path="document_change_record.yaml"
addDocChangeTab(yaml_path, table_out="kable")
}

