% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVersion.R
\name{addVersion}
\alias{addVersion}
\title{Create version string from Git}
\usage{
addVersion(x_y = "0-0", simple = T)
}
\arguments{
\item{x-y}{The major version}
}
\value{
A string representing the unique version string
}
\description{
Versioning follows this specification http://programmers.stackexchange.com/questions/141973/how-do-you-achieve-a-numeric-versioning-scheme-with-git
}
\examples{
In a directory with Git version control
addVersion()
}

