% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asanaAPI.R
\name{asanaAPI}
\alias{asanaAPI}
\alias{asana_GET}
\alias{asana_auth}
\alias{asana_check}
\alias{asana_get_projects}
\alias{asana_get_sections}
\alias{asana_get_tags}
\alias{asana_get_task_info}
\alias{asana_get_tasks}
\alias{asana_get_tasks_per_project}
\alias{asana_get_workspaces}
\alias{asana_parse}
\alias{asana_pat}
\alias{has_pat}
\title{Interact with Asana API}
\usage{
asana_GET(path, ..., pat = asana_pat())

asana_auth(pat = asana_pat())

asana_check(req)

asana_parse(req, simplifyVector = FALSE)

asana_pat()

has_pat()

asana_get_workspaces()

asana_get_projects(workspace_name)

asana_get_sections(project_name, workspace_name)

asana_get_task_info(task_id)

asana_get_tags(task_id)

asana_get_tasks(project_id)

asana_get_tasks_per_project(workspace_name, file_path = NULL)
}
\arguments{
\item{path}{Pathof API query}
}
\description{
Interact with Asana API
}

