% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromdftogantt.R
\name{fromdftogantt}
\alias{fromdftogantt}
\title{Make a Gantt diagram from a dataframe}
\usage{
fromdftogantt(df, Title = "", filename = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{Title}{Title of  diagram}

\item{filename}{Optionally write out to a html file. Default is not.}
}
\value{
A htmlwigets object representing a Gantt diagram
}
\description{
Make a Gantt diagram from a dataframe
}
\examples{
\dontrun{
# file_path <- "asana_export1.csv"
# project_start <- "2016-06-01"
# df <- parseAsana(file_path,project_start)
# df <- df[df$type \%in\% c(" Milestone", "Task"),]
# df[grep("Milestone", df$type),'type'] <- "crit"
# df[grep("Subtask", df$type),'type'] <- "done"
# df[grep("Task", df$type),'type'] <- "active"
# fromdftogantt(df)
}
}

