% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertFig.R
\name{insertFig}
\alias{insertFig}
\title{Create a Rmarkdown figure entry}
\usage{
insertFig(label, figure_list)
}
\arguments{
\item{label}{The label of the figure}

\item{figure_list}{A YAML list of figures}
}
\value{
Rmarkdown defining a kfigr chunk and a figure with details selected 
the YAML figure_list
}
\description{
Create a Rmarkdown figure entry
}
\examples{
#figure_list <- "config/figure_list.yaml"
#label <- "test_figure"
#insertFig(label, figure_list)
}

