% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAsana.R
\name{parseAsana}
\alias{parseAsana}
\title{Parse an Asana task list}
\usage{
parseAsana(file_path, project_start, format = "csv")
}
\arguments{
\item{file_path}{File exported as CSV or JSON (not implemented) from Asana.}

\item{project_start}{ISO 8601 date of project start.}
}
\value{
A dataframe suitable for producing a Gantt diagram.
}
\description{
Export an Asana task list and convert it to a dataframe ready for making a
Gantt diagram using DiagrammeR. As Asana does not have a 'Start Date' field
Tags are used to indicate periods of time. For example 'start: 0' 'end: 6'
represents a task that starts at month 0 and ends month 6. Tags are also
used to indicate 'Milestone's (important events) and 'Deliverable's (outputs
to be delivered by project). Sections are also selected...
}
\examples{
# file_path <- "asana_export1.csv"
# project_start <- "2016-06-01"
# df <- parseAsana(file_path,project_start)
}

