% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.R
\name{proposal}
\alias{proposal}
\title{A generic template for an academic proposal}
\usage{
proposal(toc = FALSE, toc_depth = 3, fig_width = 5, fig_height = 4,
  fig_caption = FALSE, highlight = "default", reference_docx = "default",
  keep_md = FALSE, md_extensions = NULL, pandoc_args = NULL)
}
\value{
MS Word (DOCX) proposal template
}
\description{
A generic template for an academic proposal
}

